+++ rl_x/algorithms/ppo/torchscript/policy.py
@@ -34,13 +34,13 @@ class Policy(nn.Module):
         single_as_shape = env.get_single_action_space_shape()
 
         self.policy_mean = nn.Sequential(
-            self.layer_init(nn.Linear(np.prod(single_os_shape).item(), nr_hidden_units)),
+            self.layer_init(nn.Linear(np.prod(single_os_shape, dtype=int).item(), nr_hidden_units)),
             nn.Tanh(),
             self.layer_init(nn.Linear(nr_hidden_units, nr_hidden_units)),
             nn.Tanh(),
-            self.layer_init(nn.Linear(nr_hidden_units, np.prod(single_as_shape).item()), std=0.01),
+            self.layer_init(nn.Linear(nr_hidden_units, np.prod(single_as_shape, dtype=int).item()), std=0.01),
         )
-        self.policy_logstd = nn.Parameter(torch.full((1, np.prod(single_as_shape).item()), np.log(std_dev).item()))
+        self.policy_logstd = nn.Parameter(torch.full((1, np.prod(single_as_shape, dtype=int).item()), np.log(std_dev).item()))
 
 
     def layer_init(self, layer, std=np.sqrt(2).item(), bias_const=(0.0)):
