+++ rl_x/algorithms/sac/pytorch/policy.py
@@ -27,13 +27,13 @@ class Policy(nn.Module):
         single_as_shape = env.get_single_action_space_shape()
 
         self.torso = nn.Sequential(
-            nn.Linear(np.prod(single_os_shape), nr_hidden_units),
+            nn.Linear(np.prod(single_os_shape, dtype=int), nr_hidden_units),
             nn.ReLU(),
             nn.Linear(nr_hidden_units, nr_hidden_units),
             nn.ReLU(),
         )
-        self.mean = nn.Linear(nr_hidden_units, np.prod(single_as_shape))
-        self.log_std = nn.Linear(nr_hidden_units, np.prod(single_as_shape))
+        self.mean = nn.Linear(nr_hidden_units, np.prod(single_as_shape, dtype=int))
+        self.log_std = nn.Linear(nr_hidden_units, np.prod(single_as_shape, dtype=int))
 
 
     def get_action(self, x):
