+++ rl_x/algorithms/sac/torchscript/policy.py
@@ -29,13 +29,13 @@ class Policy(nn.Module):
         single_as_shape = env.get_single_action_space_shape()
 
         self.torso = nn.Sequential(
-            nn.Linear(np.prod(single_os_shape).item(), nr_hidden_units),
+            nn.Linear(np.prod(single_os_shape, dtype=int).item(), nr_hidden_units),
             nn.ReLU(),
             nn.Linear(nr_hidden_units, nr_hidden_units),
             nn.ReLU(),
         )
-        self.mean = nn.Linear(nr_hidden_units, np.prod(single_as_shape).item())
-        self.log_std = nn.Linear(nr_hidden_units, np.prod(single_as_shape).item())
+        self.mean = nn.Linear(nr_hidden_units, np.prod(single_as_shape, dtype=int).item())
+        self.log_std = nn.Linear(nr_hidden_units, np.prod(single_as_shape, dtype=int).item())
 
 
     @torch.jit.export
